requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.30.0,<1.0.0
alembic>=1.14.0,<2.0.0
pydantic-settings>=2.0.0,<3.0.0
email-validator>=2.0.0,<3.0.0
PyJWT>=2.8.0,<3.0.0
argon2-cffi>=23.0.0,<24.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
httpx>=0.27.0,<1.0.0
